update the form with clearing submitting and display logic
so the goal was to:
not show reading words unless 'both' RCT was not accurate.
not show decoding unless 'both' RW was not accurate
not show phonemic until the lowest level of the first three was not default
not show Language until 'both' PA was not default
not enable the submit until LA was not default.
hidden dropdowns were to be cleared....